# apps/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Boilerplate API",
    "description": "API do boilerplate",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/clientes": {
      "post": {
        "summary": "Cadastra um cliente (já cria o checklist de onboarding)",
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ativo",
                      "onboarding",
                      "em_risco",
                      "inativo"
                    ]
                  },
                  "plano": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "growth",
                          "profissional",
                          "business"
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "cicloPlano": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "mensal",
                          "anual"
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "funisContratados": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "nome",
                  "email",
                  "status",
                  "plano",
                  "cicloPlano",
                  "funisContratados"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "nome": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "ativo",
                        "onboarding",
                        "em_risco",
                        "inativo"
                      ]
                    },
                    "plano": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "growth",
                            "profissional",
                            "business"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "cicloPlano": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "mensal",
                            "anual"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "funisContratados": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "id",
                    "nome",
                    "email",
                    "status",
                    "plano",
                    "cicloPlano",
                    "funisContratados"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/funis": {
      "post": {
        "summary": "Cadastra um funil",
        "tags": [
          "Funis"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clienteId": {
                    "type": "string"
                  },
                  "nome": {
                    "type": "string"
                  },
                  "etapa": {
                    "type": "string",
                    "enum": [
                      "onboarding",
                      "briefing_diagnostico",
                      "estrategia",
                      "copy",
                      "design",
                      "construcao_funil",
                      "pipeline",
                      "automacoes",
                      "testes_revisao",
                      "publicacao"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "nao_iniciado",
                      "em_andamento",
                      "aguardando_cliente",
                      "em_revisao",
                      "bloqueado",
                      "concluido"
                    ]
                  },
                  "responsaveis": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "dataEntrega": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "clienteId",
                  "nome",
                  "etapa",
                  "status",
                  "responsaveis",
                  "dataEntrega"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "clienteId": {
                      "type": "string"
                    },
                    "nome": {
                      "type": "string"
                    },
                    "etapa": {
                      "type": "string",
                      "enum": [
                        "onboarding",
                        "briefing_diagnostico",
                        "estrategia",
                        "copy",
                        "design",
                        "construcao_funil",
                        "pipeline",
                        "automacoes",
                        "testes_revisao",
                        "publicacao"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "nao_iniciado",
                        "em_andamento",
                        "aguardando_cliente",
                        "em_revisao",
                        "bloqueado",
                        "concluido"
                      ]
                    },
                    "responsaveis": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "dataEntrega": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "clienteId",
                    "nome",
                    "etapa",
                    "status",
                    "responsaveis",
                    "dataEntrega"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}